refactor : 폴더 및 네이밍 수정
