+++ app/join/funnels/step1/Step1Component.server.tsx
@@ -1,12 +1,12 @@
 import CheckboxSection from './components/CheckboxSection.server';
 import FormSection from './components/FormSection.client';
-import AuthTitleTextMessage from '../JoinTitleTextMessage.server';
+import JoinTitleTextMessage from '../../components/JoinTitleTextMessage.server';
 import { Spacing } from '@/components/common/Spacing';
 
 export default function Step1Component() {
   return (
     <main>
-      <AuthTitleTextMessage text={`휴대폰 번호를\n인증해주세요`} />
+      <JoinTitleTextMessage text={`휴대폰 번호를\n인증해주세요`} />
       <FormSection />
       <Spacing size={18} />
       <CheckboxSection />
