+++ app/join/funnels/step1/components/NumberForm.client.tsx
@@ -1,4 +1,4 @@
-import { useJoinContext } from '../../JoinContext';
+import { useJoinContext } from '../../../components/JoinContext';
 import { type SignUpState, useSMSMutation } from '@/apis/auth';
 import { Button } from '@/components/common/Button';
 import { Input } from '@/components/common/Input';
