+++ app/join/funnels/step3/components/FormSection.client.tsx
@@ -2,7 +2,7 @@
 
 import CertificationForm from './CertificationForm.client';
 import EmailForm from './EmailForm.client';
-import { useJoinContext } from '../../JoinContext';
+import { useJoinContext } from '../../../components/JoinContext';
 import { useFunnelContext } from '../../JoinFunnel';
 import { useEmailMutation } from '@/apis/auth';
 import BottomFixedDiv from '@/components/common/BottomFixedDiv';
